feat(): Hackathon tabs implemented
