+++ package.json
@@ -12,6 +12,7 @@
     "@chakra-ui/react": "^2.6.1",
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
+    "@types/react-tabs": "^5.0.5",
     "axios": "^1.3.4",
     "d3": "^7.8.5",
     "formik": "^2.2.9",
@@ -25,6 +26,7 @@
     "react-router-dom": "^6.9.0",
     "react-select": "^5.7.3",
     "react-spinners": "^0.13.8",
+    "react-tabs": "^6.0.1",
     "react-toastify": "^9.1.3",
     "recharts": "^2.6.2",
     "sass": "^1.62.1",
